fix: critical nonce reuse, silent drops, panics, admin auth
BREAKING: wire protocol header changed from 5 to 11 bytes (seq u16->u64)

- Fix nonce reuse vulnerability: widen sequence number from u16 to u64
  preventing ChaCha20-Poly1305 nonce repetition after 65536 packets
- Fix silent packet drops: log warn! on channel full in deliver()
- Fix decode_key() panic: return Result instead of .expect()
- Add optional bearer token auth for admin /status endpoint
- Add doc comments on 8 public types
- Clean clippy warnings (collapsible if-let)

+++ benches/relay_bench.rs
@@ -129,9 +129,9 @@ fn bench_full_pipeline(c: &mut Criterion) {
 
             // Encrypt each shard
             for (i, shard) in shards.iter().enumerate() {
-                let ct = crypto.encrypt(i as u16, shard);
+                let ct = crypto.encrypt(i as u64, shard);
                 let mut frame = vec![0u8; wire::HEADER_SIZE + ct.len()];
-                wire::encode_header(&mut frame, wire::PACKET_DATA, i as u16, ct.len() as u16);
+                wire::encode_header(&mut frame, wire::PACKET_DATA, i as u64, ct.len() as u16);
                 frame[wire::HEADER_SIZE..].copy_from_slice(&ct);
                 black_box(&frame);
             }
